backend/pyproject.toml
[project]
name = "colony-backend"
version = "0.1.0"
description = "Colony Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.12.12",
]

[tool.ruff]
line-length = 88
target-version = "py313"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "migrations",  # If using database migrations
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "C90",  # mccabe complexity
    "N",    # pep8-naming
    "D",    # pydocstyle
    "S",    # bandit security
    "A",    # flake8-builtins
    "ANN",  # flake8-annotations (ADD THIS)
    "COM",  # flake8-commas
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SLF",  # flake8-self
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "PL",   # pylint
    "RUF",  # ruff-specific rules
]

ignore = [
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D107",    # Missing docstring in __init__
    "D203",    # 1 blank line required before class docstring (conflicts with D211)
    "D213",    # Multi-line docstring summary should start at the second line
    "S101",    # Use of assert detected (common in tests)
    "PLR0913", # Too many arguments to function call
    "COM812",  # Trailing comma missing (conflicts with formatter)
    "ISC001",  # Implicitly concatenated string literals (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
# Test files can be more relaxed
"tests/**/*.py" = [
    "D",       # No docstring requirements in tests
    "S101",    # Allow assert statements in tests
    "ARG",     # Unused arguments common in test fixtures
    "PLR2004", # Magic value comparisons
]

# API route handlers often have many arguments
"*/routes/*.py" = ["PLR0913"]

# CLI and scripts can have different patterns
"scripts/*.py" = [
    "T201",    # Allow print statements
    "PLR0912", # Too many branches
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["colony", "app"]
split-on-trailing-comma = true
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 5
max-branches = 12
max-returns = 6
max-statements = 50

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
